pyproject.toml
[project]
name = "jumpstarter-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jumpstarter-all",
    "robotframework>=7.2.2",
]

[tool.uv.sources]
jumpstarter-all = { git = "https://github.com/jumpstarter-dev/jumpstarter", subdirectory = "packages/jumpstarter-all" }
